TrainAI patch migration to Harmony
