+++ TLM/TLM/Manager/Impl/GameConnectionManager.cs
@@ -3,6 +3,7 @@ namespace TrafficManager.Manager.Impl {
     using API.Manager.Connections;
     using Patch._VehicleAI.Connection;
     using Patch._VehicleAI._PassengerCarAI.Connection;
+    using Patch._VehicleAI._TrainAI.Connection;
 
     internal class GameConnectionManager: IGameConnectionManager {
 
@@ -14,9 +15,11 @@ namespace TrafficManager.Manager.Impl {
         GameConnectionManager() {
             PassengerCarAIConnection = PassengerCarAIHook.GetConnection();
             VehicleAIConnection = VehicleAIHook.GetConnection();
+            TrainAIConnection = TrainAIHook.GetConnection();
         }
 
         public IPassengerCarAIConnection PassengerCarAIConnection { get; }
         public IVehicleAIConnection VehicleAIConnection { get; }
+        public ITrainAIConnection TrainAIConnection { get; }
     }
 }
