+++ TLM/TLM/TLM.csproj
@@ -172,6 +172,13 @@
     <Compile Include="Patch\_VehicleAI\_PassengerCarAI\SimulationStepPatch.cs" />
     <Compile Include="Patch\_VehicleAI\_PassengerCarAI\StartPathFindPatch.cs" />
     <Compile Include="Patch\_VehicleAI\_PassengerCarAI\UpdateParkedVehiclePatch.cs" />
+    <Compile Include="Patch\_VehicleAI\_TrainAI\CalculateSegmentPositionPatch.cs" />
+    <Compile Include="Patch\_VehicleAI\_TrainAI\CheckNextLanePatch.cs" />
+    <Compile Include="Patch\_VehicleAI\_TrainAI\Connection\TrainAIConnection.cs" />
+    <Compile Include="Patch\_VehicleAI\_TrainAI\Connection\TrainAIHook.cs" />
+    <Compile Include="Patch\_VehicleAI\_TrainAI\ForceTrafficLightsPatch.cs" />
+    <Compile Include="Patch\_VehicleAI\_TrainAI\SimulationStep2Patch.cs" />
+    <Compile Include="Patch\_VehicleAI\_TrainAI\SimulationStepPatch.cs" />
     <Compile Include="Patch\_VehicleAI\_TramBaseAI\CalculateSegmentPositionPatch.cs" />
     <Compile Include="Patch\_VehicleAI\_TramBaseAI\CalculateSegmentPositionPatch2.cs" />
     <Compile Include="Patch\_VehicleAI\_TramBaseAI\SimulationStepPatch.cs" />
@@ -282,7 +289,6 @@
     <Compile Include="Manager\Impl\ExtVehicleManager.cs" />
     <Compile Include="Manager\Impl\VehicleRestrictionsManager.cs" />
     <Compile Include="Custom\AI\CustomHumanAI.cs" />
-    <Compile Include="Custom\AI\CustomTrainAI.cs" />
     <Compile Include="Custom\PathFinding\CustomPathManager.cs" />
     <Compile Include="LoadingExtension.cs" />
     <Compile Include="Traffic\ExtVehicleType.cs" />
