--- TLM/TMPE.API/Manager/Connections/ITramAIConnection.cs
@@ -1,6 +1,0 @@
-namespace TrafficManager.API.Manager.Connections {
-
-    public interface ITramAIConnection {
-        // in-progress
-    }
-}
